CD - working correctly

+++ python_prototype/CepstralDistance/cepsdist.py
@@ -1,10 +1,11 @@
-#Calculation of Cepstral Distance, written and tested according to matlab implementation
+# Calculation of Cepstral Distance, written and tested according to matlab implementation
 # assumption: input is horizontal, x and y are one dimensional
 
 from math import sqrt
 from realceps import realceps
 import numpy as np
 import numpy.matlib
+
 
 def cepsdist(x, y, fs):
     # param values:
@@ -31,19 +32,18 @@ def cepsdist(x, y, fs):
     num_sample = len(x)
     num_frame = np.fix((num_sample - frame + shift) / shift)
 
-    # Break up the signals into frames
-    # Not sure how to make windowing, now this part is omitted, but has to be implemented
+    # Break up the signals into frames.
     win = np.hanning(frame)
 
     idx = numpy.matlib.repmat(np.vstack(range(1, int(frame + 1))), 1, int(num_frame)) + \
           numpy.matlib.repmat(range(0, int(num_frame)) * shift, int(frame), 1)
 
-    idx=idx.astype(int)
-    x_idx=x[idx]
-    y_idx=y[idx]
+    idx = idx.astype(int)
+    x_idx = x[idx]
+    y_idx = y[idx]
 
-    X = x_idx*win[:,np.newaxis]
-    Y = y_idx*win[:,np.newaxis]
+    X = x_idx * win[:, np.newaxis]
+    Y = y_idx * win[:, np.newaxis]
 
     # Apply the cepstrum analysis
     X = np.vstack(X)  # vertical
@@ -60,12 +60,12 @@ def cepsdist(x, y, fs):
     if param_cmn == "y":
         mean_ceps_x = np.vstack(np.mean(ceps_x, 1))
         mean_ceps_y = np.vstack(np.mean(ceps_y, 1))
-        ceps_x = ceps_x-mean_ceps_x
+        ceps_x = ceps_x - mean_ceps_x
         ceps_y = ceps_y - mean_ceps_y
 
     # Calculate the cepstral distances
     err = np.square(ceps_x - ceps_y)
-    ds = 10 / np.log(10) *np.sqrt((2 * np.sum(err[1: len(err)],axis=0))+np.hstack(err[0]))
+    ds = 10 / np.log(10) * np.sqrt((2 * np.sum(err[1: len(err)], axis=0)) + np.hstack(err[0]))
 
     for i in range(len(ds)):  # min, max
         if ds[i] > 10:
